renamed form.php to better describe what page it was for added forms directory and added all the page forms into it
